Yay. GUI starts up with python 3.9
one of the bigger problems seemed to be that python 2.7 allowed
comparison between None and floats, using None as `float("-inf")`.

This isn't supported anymore in Python 3, so yeah

Also file readers output strings if not stated explicitly in Python 3.
This didn't seemed to be the case in 2.7.

Disclaimer: The GUI runs, but there are still a __lot__ of errors thrown
about

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>