Fix: no redrawing on Linux
Previously changes in the graph weren't applied until a window redraw
was forced (resize, back in foreground, ...). This was due to the
application not marking the window to refresh after the plot was
redrawn. On Windows this doesn't seem to be an issue (no idea why).

With this fix, even on Linux one can now drag and drop the control
points in the plots and see the changes being applied.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>